Changes in classes and cpp files
Made a new class "Round" that keeps track of what happens through the game. Split the 2 classes in different cpp files. These 2 .cpp files are merged in 1 .h file that then gets included in the main porogram. Gets input from the Players.txt file that contains the players and their roles. At the end of the game it creates an Output file that contains the summary of the game and what happend in each round.